feat: change folders structure for components, add custom hook for MovieCard
