+++ src/components/movieCard/styles/movieCard.scss
@@ -25,6 +25,7 @@ $color-light-grey: rgba(224, 228, 238, 0.6);
     }
   }
 
+
   &-vote {
     position: absolute;
     top: 3%;
